Add upload subcommand
The upload subcommand allows uploading ebooks in the library to a
connected device. Currently, the only unimplemented functionality is the
device-specific upload function (for the Libra 2, the only currently
supported device). Implementing `upload_ebook` for each device should be
relatively straightforward, involving only copying the file from the
source to the destination. However, I haven't decided what the naming
scheme should look like when the book is copied to the device.