Add IOKit-based USB detection
Parsing the output of `system_profiler SPUSBDataType -xml` sadly seems
to be insufficient. The Kobo Libra 2 I am using for testing stopped
displaying its mount point in the output, and so a more robust
methodology was needed. IOKit provides system APIs to iterate USB
devices and determine where they are mounted. This commit is the first
step in moving to IOKit-based USB detection for Darwin platforms
(macOS). The commit adds functionality to read the vendor and product
IDs of USB devices, as well as their BSD name. The next piece of
functionality is to use `getfsstat` to map from the BSD name to the
mount point.